rename constants directory to configuration
